relocated all methods from Calculator to CalculatorOperations
